Refactored isIllustrated.
