+++ import/index_java/src/org/vufind/index/VuFindIndexer.java
@@ -366,78 +366,6 @@ public class VuFindIndexer extends SolrIndexer
     }
 
     /**
-     * Determine if a record is illustrated.
-     *
-     * @param  LC call number
-     * @return "Illustrated" or "Not Illustrated"
-     */
-    public String isIllustrated(Record record) {
-        String leader = record.getLeader().toString();
-
-        // Does the leader indicate this is a "language material" that might have extra
-        // illustration details in the fixed fields?
-        if (leader.charAt(6) == 'a') {
-            String currentCode = "";         // for use in loops below
-
-            // List of 008/18-21 codes that indicate illustrations:
-            String illusCodes = "abcdefghijklmop";
-
-            // Check the illustration characters of the 008:
-            ControlField fixedField = (ControlField) record.getVariableField("008");
-            if (fixedField != null) {
-                String fixedFieldText = fixedField.getData().toLowerCase();
-                for (int i = 18; i <= 21; i++) {
-                    if (i < fixedFieldText.length()) {
-                        currentCode = fixedFieldText.substring(i, i + 1);
-                        if (illusCodes.contains(currentCode)) {
-                            return "Illustrated";
-                        }
-                    }
-                }
-            }
-
-            // Now check if any 006 fields apply:
-            List<VariableField> fields = record.getVariableFields("006");
-            Iterator<VariableField> fieldsIter = fields.iterator();
-            if (fields != null) {
-                while(fieldsIter.hasNext()) {
-                    fixedField = (ControlField) fieldsIter.next();
-                    String fixedFieldText = fixedField.getData().toLowerCase();
-                    for (int i = 1; i <= 4; i++) {
-                         if (i < fixedFieldText.length()) {
-                            currentCode = fixedFieldText.substring(i, i + 1);
-                            if (illusCodes.contains(currentCode)) {
-                                return "Illustrated";
-                            }
-                        }
-                    }
-                }
-            }
-        }
-
-        // Now check for interesting strings in 300 subfield b:
-        List<VariableField> fields = record.getVariableFields("300");
-        Iterator<VariableField> fieldsIter = fields.iterator();
-        if (fields != null) {
-            DataField physical;
-            while(fieldsIter.hasNext()) {
-                physical = (DataField) fieldsIter.next();
-                List<Subfield> subfields = physical.getSubfields('b');
-                for (Subfield sf: subfields) {
-                    String desc = sf.getData().toLowerCase();
-                    if (desc.contains("ill.") || desc.contains("illus.")) {
-                        return "Illustrated";
-                    }
-                }
-            }
-        }
-
-        // If we made it this far, we found no sign of illustrations:
-        return "Not Illustrated";
-    }
-
-
-    /**
      * Normalize LC numbers for sorting purposes (use only the first valid number!).
      * Will return first call number found if none pass validation,
      * or empty string if no call numbers.
