--- import/index_scripts/illustrated.bsh
@@ -1,86 +1,0 @@
-/**
- * Custom isIllustrated() script.
- *
- * This can be used to override built-in SolrMarc custom functions.  If you change
- * this script, you will need to activate it in import/marc_local.properties before
- * it will be applied during indexing.
- */
-import org.marc4j.marc.Record;
-import org.marc4j.marc.ControlField;
-import org.marc4j.marc.DataField;
-
-/**
- * Determine if a record is illustrated.
- *
- * @param  Record          record
- * @return String   "Illustrated" or "Not Illustrated"
- */
-public String isIllustrated(Record record) {
-    String leader = record.getLeader().toString();
-    
-    // Does the leader indicate this is a "language material" that might have extra
-    // illustration details in the fixed fields?
-    if (leader.charAt(6) == 'a') {
-        String currentCode = "";         // for use in loops below
-        
-        // List of 008/18-21 codes that indicate illustrations:
-        String illusCodes = "abcdefghijklmop";
-        
-        // Check the illustration characters of the 008:
-        ControlField fixedField = (ControlField) record.getVariableField("008");
-        if (fixedField != null) {
-            String fixedFieldText = fixedField.getData().toLowerCase();
-            for (int i = 18; i <= 21; i++) {
-                if (i < fixedFieldText.length()) {
-                    currentCode = fixedFieldText.substring(i, i + 1);
-                    if (illusCodes.contains(currentCode)) {
-                        return "Illustrated";
-                    }
-                }
-            }
-        }
-        
-        // Now check if any 006 fields apply:
-        List fields = record.getVariableFields("006");
-        Iterator fieldsIter = fields.iterator();
-        if (fields != null) {
-            ControlField formatField;
-            while(fieldsIter.hasNext()) {
-                fixedField = (ControlField) fieldsIter.next();
-                String fixedFieldText = fixedField.getData().toLowerCase();
-                for (int i = 1; i <= 4; i++) {
-                     if (i < fixedFieldText.length()) {
-                        currentCode = fixedFieldText.substring(i, i + 1);
-                        if (illusCodes.contains(currentCode)) {
-                            return "Illustrated";
-                        }
-                    }
-                }
-            }
-        }
-    }
-    
-    // Now check for interesting strings in 300 subfield b:
-    List fields = record.getVariableFields("300");
-    Iterator fieldsIter = fields.iterator();
-    if (fields != null) {
-        DataField physical;
-        while(fieldsIter.hasNext()) {
-            physical = (DataField) fieldsIter.next();
-            List subfields = physical.getSubfields('b');
-            Iterator subfieldsIter = subfields.iterator();
-            if (subfields != null) {
-                String desc;
-                while (subfieldsIter.hasNext()) {
-                    desc = subfieldsIter.next().getData().toLowerCase();
-                    if (desc.contains("ill.") || desc.contains("illus.")) {
-                        return "Illustrated";
-                    }
-                }
-            }
-        }
-    }
-
-    // If we made it this far, we found no sign of illustrations:
-    return "Not Illustrated";
-}
